Reorganizes Escrow contract specific code
Move escrow smart contract specific serialization code to the spefic
escrow smart contract file.

This required some exports to be made that are some interfaces and
values shared between the serializer and the contract.

Also exported the escrow specific structure and the new smart contract
stuff out of the package.